requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==4.2.27
djangorestframework==3.16.1
mysqlclient==2.2.7
pycparser==2.23
PyMySQL==1.1.2
sqlparse==0.5.4
typing_extensions==4.15.0
